Extract uploadLogWithGhUploadLog to separate module (fix line limit)
- Create src/log-upload.lib.mjs with the gh-upload-log wrapper function
- Import and re-export the function from github.lib.mjs
- Keep github.lib.mjs under 1500 line limit (now 1490 lines)
